Item edit mode test (#59)
* font size adjustments

* padding adjustments

* updates to "completed" styling

I couldn't find a scenario with the use of a line separator that I liked, the heading is better.

* made new item form text more generic

* test fixes

* attempting an "isEditable" state to toggle different views

Getting some weird behaviours when item should lose focus.

* remove unnecessary style code

* test fix

* lint fixes

* new edit mode method (thanks sam!)

I found that the delay needed to be a certain length for this to work. All functions fine but not sure if there are ways of minimising the time

* fixed test

This feels like a bit of a cheat. The test isn't able to find the delete button in the current version, but if the element exists even with the class display none, it can be found

Co-authored-by: LAPTOP-R2LRJ8HV\josep <joseph@dhproductions.co.uk>