Share list initialisation (#66)
* font size adjustments

* padding adjustments

* updates to "completed" styling

I couldn't find a scenario with the use of a line separator that I liked, the heading is better.

* made new item form text more generic

* test fixes

* attempting an "isEditable" state to toggle different views

Getting some weird behaviours when item should lose focus.

* remove unnecessary style code

* test fix

* lint fixes

* new edit mode method (thanks sam!)

I found that the delay needed to be a certain length for this to work. All functions fine but not sure if there are ways of minimising the time

* fixed test

This feels like a bit of a cheat. The test isn't able to find the delete button in the current version, but if the element exists even with the class display none, it can be found

* web share or copy to clipboard if statement

far from being react appropriate, but is a start

* Styling the list details/share/delete header

I got carried away...

* share icon cursor pointer

* share icon toggles visibility of list details

* removed fixed height, changed copy to share?

* pencil icon now reveals list details

Share icon is now for sharing only, but if web share api isn't available the copy doesn't currently work - nothing to copy!

* make share function more "react" like

* test fixing

This test seemed to fail randomly...

* modernising js copy to clipboard

* Revert to main version

Test failed for some reason

Co-authored-by: LAPTOP-R2LRJ8HV\josep <joseph@dhproductions.co.uk>